+++ programs/native/rewards/Cargo.toml
@@ -19,7 +19,6 @@ solana-sdk = { path = "../../../sdk", version = "0.12.0" }
 solana-rewards-api = { path = "../rewards_api", version = "0.12.0" }
 
 [dev-dependencies]
-solana-vote-program = { path = "../vote", version = "0.12.0" }
 solana-runtime = { path = "../../../runtime", version = "0.12.0" }
 
 [lib]
