+++ programs/native/vote/Cargo.toml
@@ -11,8 +11,6 @@ edition = "2018"
 [dependencies]
 bincode = "1.1.1"
 log = "0.4.2"
-serde = "1.0.87"
-serde_derive = "1.0.87"
 solana-logger = { path = "../../../logger", version = "0.12.0" }
 solana-metrics = { path = "../../../metrics", version = "0.12.0" }
 solana-sdk = { path = "../../../sdk", version = "0.12.0" }
@@ -22,5 +20,5 @@ solana-runtime = { path = "../../../runtime", version = "0.12.0" }
 
 [lib]
 name = "solana_vote_program"
-crate-type = ["cdylib", "lib"]
+crate-type = ["cdylib"]
