+++ programs/native/vote/src/lib.rs
@@ -3,46 +3,11 @@
 
 use bincode::deserialize;
 use log::*;
-use solana_sdk::account::{Account, KeyedAccount};
+use solana_sdk::account::KeyedAccount;
 use solana_sdk::native_program::ProgramError;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::solana_entrypoint;
-use solana_sdk::vote_program::{self, Vote, VoteInstruction, VoteState};
-
-// TODO: Deprecate the RegisterAccount instruction and its awkward delegation
-// semantics.
-fn register(keyed_accounts: &mut [KeyedAccount]) -> Result<(), ProgramError> {
-    if !vote_program::check_id(&keyed_accounts[1].account.owner) {
-        error!("account[1] is not assigned to the VOTE_PROGRAM");
-        Err(ProgramError::InvalidArgument)?;
-    }
-
-    // TODO: This assumes keyed_accounts[0] is the SystemInstruction::CreateAccount
-    // that created keyed_accounts[1]. Putting any other signed instruction in
-    // keyed_accounts[0] would allow the owner to highjack the vote account and
-    // insert itself into the leader rotation.
-    let from_id = keyed_accounts[0].signer_key().unwrap();
-
-    // Awkwardly configure the voting account to claim that the account that
-    // initially funded it is both the identity of the staker and the fullnode
-    // that should sign blocks on behalf of the staker.
-    let vote_state = VoteState::new(*from_id, *from_id);
-    vote_state.serialize(&mut keyed_accounts[1].account.userdata)?;
-
-    Ok(())
-}
-
-fn process_vote(keyed_accounts: &mut [KeyedAccount], vote: Vote) -> Result<(), ProgramError> {
-    if !vote_program::check_id(&keyed_accounts[0].account.owner) {
-        error!("account[0] is not assigned to the VOTE_PROGRAM");
-        Err(ProgramError::InvalidArgument)?;
-    }
-
-    let mut vote_state = VoteState::deserialize(&keyed_accounts[0].account.userdata)?;
-    vote_state.process_vote(vote);
-    vote_state.serialize(&mut keyed_accounts[0].account.userdata)?;
-    Ok(())
-}
+use solana_sdk::vote_program::{self, VoteInstruction};
 
 solana_entrypoint!(entrypoint);
 fn entrypoint(
@@ -63,7 +28,7 @@ fn entrypoint(
     }
 
     match deserialize(data).map_err(|_| ProgramError::InvalidUserdata)? {
-        VoteInstruction::RegisterAccount => register(keyed_accounts),
+        VoteInstruction::RegisterAccount => vote_program::register(keyed_accounts),
         VoteInstruction::Vote(vote) => {
             debug!("{:?} by {}", vote, keyed_accounts[0].signer_key().unwrap());
             solana_metrics::submit(
@@ -71,85 +36,7 @@ fn entrypoint(
                     .add_field("count", solana_metrics::influxdb::Value::Integer(1))
                     .to_owned(),
             );
-            process_vote(keyed_accounts, vote)
+            vote_program::process_vote(keyed_accounts, vote)
         }
-    }
-}
-
-pub fn create_vote_account(tokens: u64) -> Account {
-    let space = vote_program::get_max_size();
-    Account::new(tokens, space, vote_program::id())
-}
-
-pub fn register_and_deserialize(
-    from_id: &Pubkey,
-    from_account: &mut Account,
-    vote_id: &Pubkey,
-    vote_account: &mut Account,
-) -> Result<VoteState, ProgramError> {
-    let mut keyed_accounts = [
-        KeyedAccount::new(from_id, true, from_account),
-        KeyedAccount::new(vote_id, false, vote_account),
-    ];
-    register(&mut keyed_accounts)?;
-    let vote_state = VoteState::deserialize(&vote_account.userdata).unwrap();
-    Ok(vote_state)
-}
-
-pub fn vote_and_deserialize(
-    vote_id: &Pubkey,
-    vote_account: &mut Account,
-    vote: Vote,
-) -> Result<VoteState, ProgramError> {
-    let mut keyed_accounts = [KeyedAccount::new(vote_id, true, vote_account)];
-    process_vote(&mut keyed_accounts, vote)?;
-    let vote_state = VoteState::deserialize(&vote_account.userdata).unwrap();
-    Ok(vote_state)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use solana_sdk::signature::{Keypair, KeypairUtil};
-
-    #[test]
-    fn test_voter_registration() {
-        let from_id = Keypair::new().pubkey();
-        let mut from_account = Account::new(100, 0, Pubkey::default());
-
-        let vote_id = Keypair::new().pubkey();
-        let mut vote_account = create_vote_account(100);
-
-        let vote_state =
-            register_and_deserialize(&from_id, &mut from_account, &vote_id, &mut vote_account)
-                .unwrap();
-        assert_eq!(vote_state.node_id, from_id);
-        assert!(vote_state.votes.is_empty());
-    }
-
-    #[test]
-    fn test_vote() {
-        let from_id = Keypair::new().pubkey();
-        let mut from_account = Account::new(100, 0, Pubkey::default());
-
-        let vote_id = Keypair::new().pubkey();
-        let mut vote_account = create_vote_account(100);
-        register_and_deserialize(&from_id, &mut from_account, &vote_id, &mut vote_account).unwrap();
-
-        let vote = Vote::new(1);
-        let vote_state = vote_and_deserialize(&vote_id, &mut vote_account, vote.clone()).unwrap();
-        assert_eq!(vote_state.votes, vec![vote]);
-        assert_eq!(vote_state.credits(), 0);
-    }
-
-    #[test]
-    fn test_vote_without_registration() {
-        let vote_id = Keypair::new().pubkey();
-        let mut vote_account = create_vote_account(100);
-
-        let vote = Vote::new(1);
-        let vote_state = vote_and_deserialize(&vote_id, &mut vote_account, vote.clone()).unwrap();
-        assert_eq!(vote_state.node_id, Pubkey::default());
-        assert_eq!(vote_state.votes, vec![vote]);
     }
 }
